arm64: configs: enable more BPF/eBPF configs
eBPF support is now on par with the android-4.4-p common kernel.

(...or rather almost on par, only missing functionality for accessing
tunnel metadata due to backporting complexities. But that isn't relevant
for Android.)

That's only a starting point to bringing it up to the level of support
needed by modern Android versions. But for now, what's there of the
current support is functional.

Signed-off-by: K9100ii <k9100ii.ki@gmail.com>